+++ migration/src/main/scala/org/thp/thehive/migration/dto/InputDashboard.scala
@@ -1,0 +1,5 @@
+package org.thp.thehive.migration.dto
+
+import org.thp.thehive.models.Dashboard
+
+case class InputDashboard(metaData: MetaData, organisation: Option[(String, Boolean)], dashboard: Dashboard)