+++ thehive/app/org/thp/thehive/controllers/v1/MonitoringCtrl.scala
@@ -29,19 +29,18 @@ class MonitoringCtrl @Inject() (
 
   def diskUsage: Action[AnyContent] =
     entrypoint("monitor disk usage")
-      .authPermittedTransaction(db, Permissions.managePlatform)(implicit request =>
-        implicit graph =>
-          for {
-            _ <- Success(())
-            locations = diskLocations.map { dl =>
-              val file = new File(dl.location)
-              Json.obj(
-                "location"   -> dl.location,
-                "freeSpace"  -> file.getFreeSpace,
-                "totalSpace" -> file.getTotalSpace
-              )
-            }
-          } yield Results.Ok(JsArray(locations))
+      .authPermitted(Permissions.managePlatform)(_ =>
+        for {
+          _ <- Success(())
+          locations = diskLocations.map { dl =>
+            val file = new File(dl.location)
+            Json.obj(
+              "location"   -> dl.location,
+              "freeSpace"  -> file.getFreeSpace,
+              "totalSpace" -> file.getTotalSpace
+            )
+          }
+        } yield Results.Ok(JsArray(locations))
       )
 
 }