+++ thehive/app/org/thp/thehive/services/ImpactStatusSrv.scala
@@ -1,26 +1,27 @@
 package org.thp.thehive.services
 
-import akka.actor.ActorRef
+import akka.actor.typed.ActorRef
 import org.thp.scalligraph.auth.AuthContext
 import org.thp.scalligraph.models.{Database, Entity}
-import org.thp.scalligraph.services.{IntegrityCheckOps, VertexSrv}
+import org.thp.scalligraph.services.{DedupCheck, EntitySelector, IntegrityCheckOps, VertexSrv}
 import org.thp.scalligraph.traversal.TraversalOps._
 import org.thp.scalligraph.traversal.{Graph, Traversal}
 import org.thp.scalligraph.{CreateError, EntityIdOrName}
 import org.thp.thehive.models.ImpactStatus
 import org.thp.thehive.services.ImpactStatusOps._
 
-import javax.inject.{Inject, Named, Singleton}
+import javax.inject.{Inject, Provider, Singleton}
 import scala.util.{Failure, Success, Try}
 
 @Singleton
-class ImpactStatusSrv @Inject() (@Named("integrity-check-actor") integrityCheckActor: ActorRef) extends VertexSrv[ImpactStatus] {
+class ImpactStatusSrv @Inject() (integrityCheckActorProvider: Provider[ActorRef[IntegrityCheck.Request]]) extends VertexSrv[ImpactStatus] {
+  lazy val integrityCheckActor: ActorRef[IntegrityCheck.Request] = integrityCheckActorProvider.get
 
   override def getByName(name: String)(implicit graph: Graph): Traversal.V[ImpactStatus] =
     startTraversal.getByName(name)
 
   override def createEntity(e: ImpactStatus)(implicit graph: Graph, authContext: AuthContext): Try[ImpactStatus with Entity] = {
-    integrityCheckActor ! EntityAdded("ImpactStatus")
+    integrityCheckActor ! IntegrityCheck.EntityAdded("ImpactStatus")
     super.createEntity(e)
   }
 
@@ -42,15 +43,4 @@ object ImpactStatusOps {
   }
 }
 
-class ImpactStatusIntegrityCheckOps @Inject() (val db: Database, val service: ImpactStatusSrv) extends IntegrityCheckOps[ImpactStatus] {
-  override def resolve(entities: Seq[ImpactStatus with Entity])(implicit graph: Graph): Try[Unit] =
-    entities match {
-      case head :: tail =>
-        tail.foreach(copyEdge(_, head))
-        service.getByIds(tail.map(_._id): _*).remove()
-        Success(())
-      case _ => Success(())
-    }
-
-  override def globalCheck(): Map[String, Int] = Map.empty
-}
+class ImpactStatusIntegrityCheck @Inject() (val db: Database, val service: ImpactStatusSrv) extends DedupCheck[ImpactStatus]