+++ thehive/app/org/thp/thehive/services/ObservableTypeSrv.scala
@@ -1,6 +1,6 @@
 package org.thp.thehive.services
 
-import akka.actor.ActorRef
+import akka.actor.typed.ActorRef
 import org.thp.scalligraph.auth.AuthContext
 import org.thp.scalligraph.models.{Database, Entity}
 import org.thp.scalligraph.services._
@@ -10,13 +10,14 @@ import org.thp.scalligraph.{BadRequestError, CreateError, EntityIdOrName}
 import org.thp.thehive.models._
 import org.thp.thehive.services.ObservableTypeOps._
 
-import javax.inject.{Inject, Named, Provider, Singleton}
+import javax.inject.{Inject, Provider, Singleton}
 import scala.util.{Failure, Success, Try}
 
 @Singleton
-class ObservableTypeSrv @Inject() (_observableSrv: Provider[ObservableSrv], @Named("integrity-check-actor") integrityCheckActor: ActorRef)
+class ObservableTypeSrv @Inject() (_observableSrv: Provider[ObservableSrv], integrityCheckActorProvider: Provider[ActorRef[IntegrityCheck.Request]])
     extends VertexSrv[ObservableType] {
-  lazy val observableSrv: ObservableSrv = _observableSrv.get
+  lazy val observableSrv: ObservableSrv                          = _observableSrv.get
+  lazy val integrityCheckActor: ActorRef[IntegrityCheck.Request] = integrityCheckActorProvider.get
 
   override def getByName(name: String)(implicit graph: Graph): Traversal.V[ObservableType] =
     startTraversal.getByName(name)
@@ -24,7 +25,7 @@ class ObservableTypeSrv @Inject() (_observableSrv: Provider[ObservableSrv], @Nam
   override def exists(e: ObservableType)(implicit graph: Graph): Boolean = startTraversal.getByName(e.name).exists
 
   override def createEntity(e: ObservableType)(implicit graph: Graph, authContext: AuthContext): Try[ObservableType with Entity] = {
-    integrityCheckActor ! EntityAdded("ObservableType")
+    integrityCheckActor ! IntegrityCheck.EntityAdded("ObservableType")
     super.createEntity(e)
   }
 
@@ -62,15 +63,4 @@ object ObservableTypeOps {
   }
 }
 
-class ObservableTypeIntegrityCheckOps @Inject() (val db: Database, val service: ObservableTypeSrv) extends IntegrityCheckOps[ObservableType] {
-  override def resolve(entities: Seq[ObservableType with Entity])(implicit graph: Graph): Try[Unit] =
-    entities match {
-      case head :: tail =>
-        tail.foreach(copyEdge(_, head))
-        service.getByIds(tail.map(_._id): _*).remove()
-        Success(())
-      case _ => Success(())
-    }
-
-  override def globalCheck(): Map[String, Int] = Map.empty
-}
+class ObservableTypeIntegrityCheck @Inject() (val db: Database, val service: ObservableTypeSrv) extends DedupCheck[ObservableType]