+++ thehive/app/org/thp/thehive/services/TagSrv.scala
@@ -1,13 +1,13 @@
 package org.thp.thehive.services
 
-import akka.actor.ActorRef
+import akka.actor.typed.ActorRef
 import org.apache.tinkerpop.gremlin.process.traversal.TextP
 import org.apache.tinkerpop.gremlin.structure.Vertex
 import org.thp.scalligraph.EntityIdOrName
 import org.thp.scalligraph.auth.AuthContext
 import org.thp.scalligraph.models.{Database, Entity}
 import org.thp.scalligraph.services.config.{ApplicationConfig, ConfigItem}
-import org.thp.scalligraph.services.{EdgeSrv, EntitySelector, IntegrityCheckOps, VertexSrv}
+import org.thp.scalligraph.services._
 import org.thp.scalligraph.traversal.TraversalOps._
 import org.thp.scalligraph.traversal.{Converter, Graph, Traversal}
 import org.thp.scalligraph.utils.FunctionalCondition.When
@@ -16,7 +16,7 @@ import org.thp.thehive.services.OrganisationOps._
 import org.thp.thehive.services.TagOps._
 
 import java.util.{Date, Map => JMap}
-import javax.inject.{Inject, Named, Provider, Singleton}
+import javax.inject.{Inject, Provider, Singleton}
 import scala.util.matching.Regex
 import scala.util.{Success, Try}
 
@@ -25,9 +25,10 @@ class TagSrv @Inject() (
     organisationSrv: OrganisationSrv,
     taxonomySrvProvider: Provider[TaxonomySrv],
     appConfig: ApplicationConfig,
-    @Named("integrity-check-actor") integrityCheckActor: ActorRef
+    integrityCheckActorProvider: Provider[ActorRef[IntegrityCheck.Request]]
 ) extends VertexSrv[Tag] {
-  lazy val taxonomySrv: TaxonomySrv = taxonomySrvProvider.get
+  lazy val taxonomySrv: TaxonomySrv                              = taxonomySrvProvider.get
+  lazy val integrityCheckActor: ActorRef[IntegrityCheck.Request] = integrityCheckActorProvider.get
 
   val taxonomyTagSrv = new EdgeSrv[TaxonomyTag, Taxonomy, Tag]
   private val freeTagColourConfig: ConfigItem[String, String] =
@@ -76,7 +77,7 @@ class TagSrv @Inject() (
     } yield tag
 
   def create(tag: Tag)(implicit graph: Graph, authContext: AuthContext): Try[Tag with Entity] = {
-    integrityCheckActor ! EntityAdded("Tag")
+    integrityCheckActor ! IntegrityCheck.EntityAdded("Tag")
     super.createEntity(tag)
   }
 
@@ -148,6 +149,8 @@ object TagOps {
         .has(_.namespace, freeTagNamespace)
     }
 
+    def freetags: Traversal.V[Tag] = traversal.has(_.namespace, TextP.startingWith("_freetag_"))
+
     def getFreetag(organisationSrv: OrganisationSrv, idOrName: EntityIdOrName)(implicit authContext: AuthContext): Traversal.V[Tag] =
       idOrName.fold(traversal.getByIds(_), traversal.has(_.predicate, _)).freetags(organisationSrv)
 
@@ -174,40 +177,12 @@ object TagOps {
   }
 }
 
-class TagIntegrityCheckOps @Inject() (val db: Database, val service: TagSrv) extends IntegrityCheckOps[Tag] {
+class TagIntegrityCheck @Inject() (val db: Database, val service: TagSrv) extends DedupCheck[Tag] with GlobalCheck[Tag] with IntegrityCheckOps[Tag] {
+  override def extraFilter(traversal: Traversal.V[Tag]): Traversal.V[Tag] =
+    traversal
+      .freetags
+      .filterNot(_.or(_.alert, _.observable, _.`case`, _.caseTemplate))
 
-  override def resolve(entities: Seq[Tag with Entity])(implicit graph: Graph): Try[Unit] = {
-    EntitySelector.firstCreatedEntity(entities).foreach {
-      case (head, tail) =>
-        tail.foreach(copyEdge(_, head))
-        val tailIds = tail.map(_._id)
-        logger.debug(s"Remove duplicated vertex: ${tailIds.mkString(",")}")
-        service.getByIds(tailIds: _*).remove()
-    }
-    Success(())
-  }
-
-  override def globalCheck(): Map[String, Int] =
-    service
-      .pagedTraversalIds(
-        db,
-        100,
-        _.filter(_.taxonomy.has(_.namespace, TextP.startingWith("_freetags_")))
-          .filterNot(_.or(_.alert, _.observable, _.`case`, _.caseTemplate, _.taxonomy))
-      ) { ids =>
-        db.tryTransaction { implicit graph =>
-          Try {
-            val orphans = service
-              .getByIds(ids: _*)
-              ._id
-              .toSeq
-            if (orphans.nonEmpty) {
-              service.getByIds(orphans: _*).remove()
-              Map("orphan" -> orphans.size)
-            } else Map.empty[String, Int]
-          }
-        }.getOrElse(Map("globalFailure" -> 1))
-      }
-      .reduceOption(_ <+> _)
-      .getOrElse(Map.empty)
+  override def globalCheck(traversal: Traversal.V[Tag])(implicit graph: Graph): Map[String, Long] =
+    Map("orphan" -> traversal.sideEffect(_.drop()).getCount)
 }